Fix formatting
